Preserve creation order when serializing
This replaces the private `_creationRuntimeIdx` and `_initialCreationTime`
properties with a public `.creationIndex` property.

Cookies now get a creation time at construction, but this is replaced
when `jar.setCookie` according to either the wall time or the `now`
option to that method.

Prior to this change, cookies were returned by MemoryCookieStore's
getAllCookies in a strange order (sorted by domain then path then key),
which breaks `cookieCompare` for cookies with creation times in the same
millisecond.

Sorting by `.creationIndex` and then removing that property will allow
the deserializer to assign new process-specific `.creationIndex` values.
But, if the underlying `Store` wants to set that upon retrieval, then
they can do that too.

By making the property non-enumerable, `assert.deepEqual` can work as
expected ;)