setCookie arg must be Cookie or string
Addresses #132, partially. The error message is clearer, indicating that
this method only accepts the afformentioned types. The part of 132 that
it doesn't address is the whole "Cookie instance from another version of
the `tough-cookie` package" thing, which isn't really solvable unless we
somehow expose the Cookie class that "belongs to" this CookieJar class,
which is maybe another patch.