Update config to support ES6 classes.
Targeting ES5 causes the "fixes issue #145" test in cookieJar.spec.ts
to fail. Changing the target to the next newest (ES6) requires adding
"moduleResolution" to the tsconfig, which then causes issues with using
`import` in our TypeScript, but `require` in the legacy vows tests. To
support both, we have to bump to a newer version of TypeScript to use
the new "node16" module/moduleResolution. (Note: I didn't jump to the
_latest_ TypeScript to try to support older projects, as well.)