[v5] Update config (#269)
* Update config to support ES6 classes.

Targeting ES5 causes the "fixes issue #145" test in cookieJar.spec.ts
to fail. Changing the target to the next newest (ES6) requires adding
"moduleResolution" to the tsconfig, which then causes issues with using
`import` in our TypeScript, but `require` in the legacy vows tests. To
support both, we have to bump to a newer version of TypeScript to use
the new "node16" module/moduleResolution. (Note: I didn't jump to the
_latest_ TypeScript to try to support older projects, as well.)

* Avoid unnecessarily running compiled test files.

* Fix flaky test.

* Avoid hacky reliance on inherited property of global object.

* Avoid needing to use `resolveJsonModule`.

* Remove config that's just the default.

See: https://huafu.github.io/ts-jest/user/config/isolatedModules

* Update tsconfig to target node 16 and only compile ./lib

* Chang `dist/lib` imports to just `dist`.

* add granularity to npm scripts