Fix allow listed `files` for what to include in npm package
When #296 was merged, the list of files to include in `package.json` was not updated. This PR fixes that.

The `.npmignore` file is also removed because I find it causes some unexpected behaviors like ignoring the contents of `.gitignore`. It's a bit of extra work to maintain the allow listed `files` but it's generally safer.