Fix regression bug in `domainMatch`
The code in `domainMatch` needs to make comparisions using the `_str` and `_domStr` variables as these are canonicalized versions of the `domain` and `cookieDomain` arguments.

One of the comparisions was mistakenly using the `cookieDomain` when it should have been using `_domStr`. This PR closes #499 by using the correct variable for the comparison and adds a test that covers this case to our list of cases for `domainMatch`.