Support publishing of both ESM and CJS
The [ESM supported backported to Node 20.19.0](https://github.com/nodejs/node/releases/tag/v20.19.0) is nice but not sufficient for supporting users on earlier versions of Node v20 which is still in active maintenance until 2026-04-30. This PR adds in dual-publishing of both ESM and CJS and ensures that entrypoints and types are correctly declared.