Support publishing of both ESM and CJS (#536)
* Support publishing of both ESM and CJS

The [ESM supported backported to Node 20.19.0](https://github.com/nodejs/node/releases/tag/v20.19.0) is nice but not sufficient for supporting users on earlier versions of Node v20 which is still in active maintenance until 2026-04-30. This PR adds in dual-publishing of both ESM and CJS and ensures that entrypoints and types are correctly declared.

* Fix api extractor path

* 6.0.0-rc.2

* Update package.json

Co-authored-by: Will Harney <62956339+wjhsf@users.noreply.github.com>

* Add validation for ESM/CJS types and entrypoints

* Add validation for ESM/CJS types and entrypoints

* Add explicit dependency on Are The Types Wrong cli + npm scripts

---------

Co-authored-by: Will Harney <62956339+wjhsf@users.noreply.github.com>